requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
Django==5.2
djangorestframework==3.16.0
minio==7.2.15
psycopg2-binary==2.9.10
pycparser==2.22
pycryptodome==3.22.0
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
